requirements.txt
praw
configparser
mal-api
re
